+++ source/extensions/isaacsim.replicator.experimental.mobility_gen/python/impl/__init__.py
@@ -18,6 +18,7 @@ from .camera import MobilityGenCamera
 from .common import Buffer, Module
 from .config import Config
 from .inputs import Gamepad, GamepadDriver, Keyboard, KeyboardDriver
+from .nurec_overrides import apply_nurec_replay_overrides, is_nurec_stage
 from .occupancy_map import OccupancyMap
 from .path_planner import compress_path, generate_paths
 from .pose_samplers import GridPoseSampler, UniformPoseSampler
@@ -54,9 +55,11 @@ __all__ = [
     "SCENARIOS",
     "SensorConfig",
     "UniformPoseSampler",
+    "apply_nurec_replay_overrides",
     "apply_sensor_overrides",
     "compress_path",
     "generate_paths",
+    "is_nurec_stage",
     "load_scenario",
     "log_camera_properties",
     "save_sensor_overrides",